LightLogR: Process Data from Wearable Light Loggers and Optical Radiation Dosimeters

Import, processing, validation, and visualization of personal light exposure measurement data from wearable devices. The package implements features such as the import of data and metadata files, conversion of common file formats, validation of light logging data, verification of crucial metadata, calculation of common parameters, and semi-automated analysis and visualization.

See also

Author

Maintainer: Johannes Zauner johannes.zauner@tum.de (ORCID)

Authors:

Manuel Spitschan manuel.spitschan@tum.de (ORCID)

Steffen Hartmeyer steffen.hartmeyer@epfl.ch (ORCID)

Other contributors:

MeLiDos [funder]

EURAMET (European Association of National Metrology Institutes. Website: www.euramet.org. Grant Number: 22NRM05 MeLiDos. Grant Statement: The project (22NRM05 MeLiDos) has received funding from the European Partnership on Metrology, co-financed from the European Union’s Horizon Europe Research and Innovation Programme and by the Participating States.) [funder]

European Union (Co-funded by the European Union. Views and opinions expressed are however those of the author(s) only and do not necessarily reflect those of the European Union or EURAMET. Neither the European Union nor the granting authority can be held responsible for them.) [funder]

TSCN-Lab (www.tscnlab.org) [copyright holder]