Create a simple plot of light logger data, facetted by Day

gg_day will create a simple ggplot for every data in a dataset. The result can further be manipulated like any ggplot. This will be sensible to refine styling or guides.

Usage

gg_day(
  dataset,
  start.date = NULL,
  end.date = NULL,
  x.axis = Time,
  y.axis,
  key = NULL,
  scales = "free_y",
  y.axis.breaks = 10^(0:5),
  y.scale.log10 = TRUE,
  x.axis.label = "Time of Day",
  y.axis.label = "Lightlevel",
  format.day = "%d/%m",
  title = NULL,
  subtitle = NULL
)

Arguments

dataset

light logger dataset. Expects a dataframe.

start.date

Choose an optional start date within your dataset. Expects a date, which can also be a character that is interpretable as a date, e.g., "2023-06-03". Defaults to NULL, which means that the plot starts with the earliest date within the dataset.

end.date

Choose an optional end date within your dataset. see start.date for additional details.

x.axis

column name that contains the timedate. Expects a symbol. Needs to be part of the dataset.

y.axis

column name that contains the lightlevel. Expects a symbol. Needs to be part of the dataset.

key

optional column name that defines separate sets. Expects a symbol. Needs to be part of the dataset. By default, the ggplot2::scale_color_viridis_d is used. This can be overwritten outside the function (see examples).

scales

For ggplot2::facet_wrap, should scales be "fixed", "free" or free in one dimension ("free_y" is the default). Expects a character.

y.axis.breaks

Where should breaks occur on the y.axis? Expects a numeric vector with all the breaks. If you want to activate the default behaviour of ggplot2, you need to put in ggplot2::waiver().

y.scale.log10

Should y be scaled on a log10 basis? Expects a logical.

x.axis.label

label for the x-axis. Expects a character.

y.axis.label

label for the y-axis. Expects a character.

format.day

Label for each day. Default is %d/%m, which shows the day and month. Expects a character. For an overview of sensible options look at strptime

title

Plot title. Expects a character.

subtitle

Plot subtitle. Expects a character.

Value

A ggplot object

Examples

filepath <- system.file("extdata/sample_data_LYS.csv", package = "LightLogR")
sampledata <- import.LYS(filepath, file.id = TRUE)
#> Error in expr({    tmp <- readr::read_csv(paste0(path, filename), n_max = n_max,         col_types = c("cfddddddddddd"), id = if (file.id)             "filepath")    tmp <- tmp %>% dplyr::rename(Datetime = timestamp) %>% dplyr::mutate(Datetime = Datetime %>%         lubridate::dmy_hms(tz = tz))}): could not find function "expr"
gg_day(sampledata, x.axis = Datetime, y.axis = mEDI, y.axis.label = "mEDI (lx)", title = "Nonvisual Stimulus for 2 Days")
#> Error in eval(expr, envir, enclos): object 'sampledata' not found

#use `key`for separation of different sets
plot <- gg_day(
sample.data.environment,
scales = "fixed",
x.axis = Datetime,
y.axis = mEDI,
y.axis.label = "mEDI (lx)",
key = Source)
#> Error in gg_day(sample.data.environment, scales = "fixed", x.axis = Datetime,     y.axis = mEDI, y.axis.label = "mEDI (lx)", key = Source): axis_columns %in% names(dataset) are not all TRUE
plot
#> function (x, y, ...) 
#> UseMethod("plot")
#> <bytecode: 0x561b0678d698>
#> <environment: namespace:base>

#you can easily overwrite the color scale afterwards
plot + ggplot2::scale_color_discrete()
#> NULL