Import Dataset from ActLumus

Imports a dataset and does the necessary transformations to get the right column formats. Unless specified otherwise, the function will set the timezone of the data to UTC.

Usage

import.ActLumus(
  filename,
  path = NULL,
  file.id = FALSE,
  n_max = Inf,
  tz = "UTC"
)

Arguments

filename

Filename(s) for the Dataset. Can also contain the filepath, but path must then be NULL. Expects a character. If the vector is longer than 1, multiple files will be read in into one Tibble.

path

Optional path for the dataset(s). NULL is the default. Expects a character.

file.id

Should the filename be stored as part of the data? Useful if you read in data from multiple files

n_max

maximum number of lines to read. Default is Inf.

Value

Tibble/Dataframe with a POSIXct column for the datetime

Examples

filepath <- system.file("extdata/sample_data_ActLumus.txt", package = "LightLogR")
sampledata <- import.ActLumus(filepath)
#> Error in expr({    tmp <- readr::read_delim(paste0(path, filename), skip = 32,         delim = ";", n_max = n_max, col_types = paste0("c", rep("d",             32)), id = if (file.id)             "filepath")    tmp <- tmp %>% dplyr::rename(Datetime = `DATE/TIME`) %>%         dplyr::mutate(Datetime = Datetime %>% lubridate::dmy_hms(tz = tz))}): could not find function "expr"
dplyr::slice_sample(sampledata, n = 5) %>% gt::gt()
#> Error in dplyr::slice_sample(sampledata, n = 5) %>% gt::gt(): could not find function "%>%"
gg_day(sampledata, x.axis = Datetime, y.axis = `MELANOPIC EDI`, y.axis.label = "mEDI (lx)")
#> Error in eval(expr, envir, enclos): object 'sampledata' not found