Import a light logger dataset or related data

Imports a dataset and does the necessary transformations to get the right column formats. Unless specified otherwise, the function will set the timezone of the data to UTC. It will also enforce an id to separate different datasets and will order/arrange the dataset within each id.

Usage

import.Dataset(device, ...)

import.ActLumus(
  filename,
  path = NULL,
  n_max = Inf,
  tz = "UTC",
  ID.colname = Id,
  auto.id = ".*",
  manual.id = NULL
)

import.LYS(
  filename,
  path = NULL,
  n_max = Inf,
  tz = "UTC",
  ID.colname = Id,
  auto.id = ".*",
  manual.id = NULL
)

Arguments

device

From what device do you want to import? For every supported device, there is a sample data file that you can use to test the function (see the examples). Currently the following devices are supported (followed by the device.ext spec to access the sample file):

"ActLumus" (ActLumus.txt)

"LYS" (LYS.csv)

...

Parameters that get handed down to the specific import functions

filename

Filename(s) for the Dataset. Can also contain the filepath, but path must then be NULL. Expects a character. If the vector is longer than 1, multiple files will be read in into one Tibble.

path

Optional path for the dataset(s). NULL is the default. Expects a character.

n_max

maximum number of lines to read. Default is Inf.

tz

Timezone of the data. "UTC" is the default. Expects a character. You can look up the supported timezones with OlsonNames().

ID.colname

Lets you specify a column for the participant id. Expects a symbol (Default is Id). This column will be used for grouping (dplyr::group_by()).

auto.id

If the Id.colname column is added to the dataset, the Id can be automatically extracted from the filename. The argument expects a regular expression regex and will by default just give the whole filename without file extension.

manual.id

If this argument is not NULL, and no ID column is part of the dataset, this character scalar will be used. Don´t use this argument if multiple files from different participants are used!.

Value

Tibble/Dataframe with a POSIXct column for the datetime

Details

If the Id column is already part of the dataset it will just use this column. If the column is not present it will add this column and fill it with the filename of the importfile (see param auto.id).

Examples

Imports made easy

The set of import functions provide a convenient way to import light logger data that is then perfectly formatted to add metadata, make visualizations and analyses. There are a number of devices supported, where import should just work out of the box. To get an overview, you can simply call the supported.devices dataset. The list will grow continuously as the package is maintained.

supported.devices
#> [1] "LYS"      "ActLumus"

To import a file, simple specify the filename (and path) and feed it to the import.Dataset function. There are sample datasets for all devices.

The import functions provide a basic overview of the data after import, such as the intervals between measurements or the start and end dates.

filepath <- system.file("extdata/sample_data_LYS.csv", package = "LightLogR")
dataset <- import.Dataset("LYS", filepath)
#> Successfully read in 11422 observations from LYS-file
#> Timezone set is UTC.
#> The system timezone is Europe/Berlin. Please correct if necessary!
#> Start: 2023-06-21 00:00:12
#> End: 2023-06-22 23:59:48
#> Timespan: 2 days
#> Observation intervals: 
#>   Id              interval.time     n pct    
#> 1 sample_data_LYS 15s           10015 87.689%
#> 2 sample_data_LYS 16s            1367 11.969%
#> 3 sample_data_LYS 17s              23 0.201% 
#> 4 sample_data_LYS 18s              16 0.140%

Import functions can also be called directly:

filepath <- system.file("extdata/sample_data_ActLumus.txt", package = "LightLogR")
dataset <- import.ActLumus(filepath)
#> Successfully read in 2065 observations from ActLumus-file
#> Timezone set is UTC.
#> The system timezone is Europe/Berlin. Please correct if necessary!
#> Start: 2023-06-23 11:57:14
#> End: 2023-07-14 22:30:30
#> Timespan: 21 days
#> Observation intervals: 
#>   Id                   interval.time              n pct  
#> 1 sample_data_ActLumus 60s (~1 minutes)        2059 100% 
#> 2 sample_data_ActLumus 132s (~2.2 minutes)        1 0%   
#> 3 sample_data_ActLumus 11324s (~3.15 hours)       1 0%   
#> 4 sample_data_ActLumus 18054s (~5.02 hours)       1 0%   
#> 5 sample_data_ActLumus 60872s (~16.91 hours)      1 0%   
#> 6 sample_data_ActLumus 1638474s (~2.71 weeks)     1 0%

dataset %>%
dplyr::select(Datetime, TEMPERATURE, LIGHT, MEDI) %>%
dplyr::slice(1500:1505) %>%
flextable::flextable() %>%
flextable::autofit()
#> Adding missing grouping variables: `Id`

Id

Datetime

TEMPERATURE

LIGHT

MEDI

sample_data_ActLumus

2023-07-14 13:04:18

29.88

21,847.64

20,660.74

sample_data_ActLumus

2023-07-14 13:05:18

29.88

22,033.77

20,834.91

sample_data_ActLumus

2023-07-14 13:06:18

29.81

21,769.50

20,600.82

sample_data_ActLumus

2023-07-14 13:07:18

29.69

21,177.20

20,061.68

sample_data_ActLumus

2023-07-14 13:08:18

29.44

20,738.98

19,662.14

sample_data_ActLumus

2023-07-14 13:09:18

29.31

20,255.17

19,203.94