Get all the supported devices in LightLogR

Returns a vector of all the supported devices in LightLogR.

Usage

supported_devices()

Value

A character vector of all supported devices

Details

These are all supported devices where there is a dedicated import function. Import functions can be called either through import_Dataset() with the respective device = "device" argument, or directly, e.g., import$ActLumus().

See also

Examples

supported_devices()
#>  [1] "ActLumus"              "ActTrust"              "Actiwatch_Spectrum"   
#>  [4] "Actiwatch_Spectrum_de" "Circadian_Eye"         "Clouclip"             
#>  [7] "DeLux"                 "GENEActiv_GGIR"        "Kronowise"            
#> [10] "LIMO"                  "LYS"                   "LiDo"                 
#> [13] "LightWatcher"          "MotionWatch8"          "OcuWEAR"              
#> [16] "Speccy"                "SpectraWear"           "VEET"                 
#> [19] "nanoLambda"