Create Datetime bins for visualization and calculation

cut_Datetime is a wrapper around lubridate::round_date() (and friends) combined with dplyr::mutate(), to create a new column in a light logger dataset with a specified binsize. This can be "3 hours", "15 secs", or "0.5 days". It is a useful step between a dataset and a visualization or summary step.

Usage

cut_Datetime(
  dataset,
  unit = "3 hours",
  type = "round",
  Datetime.colname = Datetime,
  New.colname = Datetime.rounded,
  group_by = FALSE,
  ...
)

Arguments

dataset

A light logger dataset. Expects a dataframe. If not imported by LightLogR, take care to choose a sensible variable for the Datetime.colname.

unit

Unit of binning. See lubridate::round_date() for examples.

type

One of "round"(the default), "ceiling" or "floor". Setting chooses the relevant function from lubridate.

Datetime.colname

column name that contains the datetime. Defaults to "Datetime" which is automatically correct for data imported with LightLogR. Expects a symbol. Needs to be part of the dataset.

New.colname

Column name for the added column in the dataset.

group_by

Should the data be grouped by the new column? Defaults to FALSE

...

Parameter handed over to lubridate::round_date() and siblings

Value

a data.frame object identical to dataset but with the added column of binned datetimes.

Examples

#compare Datetime and Datetime.rounded
sample.data.environment %>%
  cut_Datetime() %>%
  dplyr::slice_sample(n = 5)
#> # A tibble: 10 × 4
#> # Groups:   Id [2]
#>    Datetime               MEDI Id          Datetime.rounded   
#>    <dttm>                <dbl> <chr>       <dttm>             
#>  1 2023-08-18 07:26:32 10365.  Environment 2023-08-18 06:00:00
#>  2 2023-08-20 12:58:32 80979.  Environment 2023-08-20 12:00:00
#>  3 2023-08-15 01:27:32     0   Environment 2023-08-15 00:00:00
#>  4 2023-08-15 07:00:02  3858.  Environment 2023-08-15 06:00:00
#>  5 2023-08-19 16:40:32 26145.  Environment 2023-08-19 18:00:00
#>  6 2023-08-18 13:34:11  1628.  Participant 2023-08-18 15:00:00
#>  7 2023-08-16 03:25:31     0   Participant 2023-08-16 03:00:00
#>  8 2023-08-18 17:45:01   726.  Participant 2023-08-18 18:00:00
#>  9 2023-08-19 17:07:21    42.4 Participant 2023-08-19 18:00:00
#> 10 2023-08-19 05:08:01     0   Participant 2023-08-19 06:00:00