Get the import expression for a device

Returns the import expression for all device in LightLogR.

Usage

ll_import_expr()

Value

A list of import expressions for all supported devices

Details

These expressions are used to import and prepare data from specific devices. The list is made explicit, so that a user, requiring slight changes to the import functions, (e.g., because a timestamp is formatted differently) can modify or add to the list. The list can be turned into a fully functional import function through import_adjustment().

See also

Examples

ll_import_expr()[1]
#> $ActLumus
#> {
#>     data <- suppressMessages(readr::read_delim(filename, skip = 32, 
#>         delim = ";", n_max = n_max, col_types = paste0("c", paste0(rep("d", 
#>             32), collapse = "")), id = "file.name", locale = locale, 
#>         name_repair = "universal", ...))
#>     data <- data %>% dplyr::rename(Datetime = DATE.TIME, MEDI = MELANOPIC.EDI) %>% 
#>         dplyr::mutate(Datetime = Datetime %>% lubridate::dmy_hms(tz = tz))
#> }
#>